Chore: Move SubmissionPage content into hashing folder
