+++ lib/features/verification/logic/VerificationLogic.dart
@@ -1,10 +1,11 @@
 import 'dart:convert';
 import 'package:chopper/chopper.dart';
-import 'package:decentproof/pages/submissionpage/logic/MessageVerificationService.dart';
 import 'package:decentproof/shared/generated/openapi.swagger.dart';
 import 'package:convert/convert.dart';
 import 'package:http_certificate_pinning/http_certificate_pinning.dart';
 import 'package:pointycastle/pointycastle.dart';
+
+import '../../hashing/logic/blockchain/MessageVerificationService.dart';
 
 class VerificationLogic {
   final MessageVerificationService verificationService;
